Refactored DatabaseController.java
