+++ modules/spring-webflux-aggregator/src/main/java/io/starter/controller/RateController.java
@@ -1,0 +1,51 @@
+package io.starter.controller;
+
+import io.starter.config.ScheduleConfig;
+import io.starter.entity.LeagueEntity;
+import io.starter.service.DataAccessService;
+import io.starter.service.NinjaDataSyncService;
+import io.starter.service.PoeNinjaService;
+
+import lombok.extern.log4j.Log4j2;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/rates")
+@Log4j2
+public class RateController {
+
+  private final DataAccessService dataAccessService;
+  private final PoeNinjaService poeNinjaService;
+  private final NinjaDataSyncService ninjaDataSyncService;
+
+  public RateController(DataAccessService dataAccessService, PoeNinjaService poeNinjaService,
+                        NinjaDataSyncService ninjaDataSyncService) {
+    this.dataAccessService = dataAccessService;
+    this.poeNinjaService = poeNinjaService;
+    this.ninjaDataSyncService = ninjaDataSyncService;
+  }
+
+  @PostMapping("/load")
+  public void loadRates() {
+    dataAccessService.findLeagues().forEach(this::loadRates);
+  }
+
+  @Scheduled(cron = ScheduleConfig.A8R_ADD_CRON)
+  public void updateRates() {
+    dataAccessService.findLeagues().forEach(league ->
+        poeNinjaService.getRates(league.getName())
+            .subscribe(response -> ninjaDataSyncService.updateCurrencies(response.getBody(), league))
+    );
+  }
+
+  private void loadRates(LeagueEntity league) {
+    poeNinjaService.getRates(league.getName())
+        .subscribe(response -> {
+          ninjaDataSyncService.loadCurrency(response.getBody(), league);
+          log.info("{} - Loaded {} Rates", league.getName(), dataAccessService.findRatesByLeague(league).size());
+        });
+  }
+}
