+++ modules/spring-webflux-aggregator/src/main/java/io/starter/controller/UniqueJewelController.java
@@ -1,0 +1,42 @@
+package io.starter.controller;
+
+import io.starter.entity.LeagueEntity;
+import io.starter.service.DataAccessService;
+import io.starter.service.NinjaDataSyncService;
+import io.starter.service.PoeNinjaService;
+
+import lombok.extern.log4j.Log4j2;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/unique-jewels")
+@Log4j2
+public class UniqueJewelController {
+
+  private final DataAccessService dataAccessService;
+  private final PoeNinjaService poeNinjaService;
+  private final NinjaDataSyncService ninjaDataSyncService;
+
+  public UniqueJewelController(DataAccessService dataAccessService,
+                               PoeNinjaService poeNinjaService,
+                               NinjaDataSyncService ninjaDataSyncService) {
+    this.dataAccessService = dataAccessService;
+    this.poeNinjaService = poeNinjaService;
+    this.ninjaDataSyncService = ninjaDataSyncService;
+  }
+
+  @PostMapping("/load")
+  public void loadUniqueJewels() {
+    dataAccessService.findLeagues().forEach(this::loadUniqueJewels);
+  }
+
+  private void loadUniqueJewels(LeagueEntity league) {
+    poeNinjaService.getUniqueJewels(league.getName())
+        .subscribe(response -> {
+          ninjaDataSyncService.loadUniqueJewels(response.getBody(), league);
+          log.info("{} - Loaded {} Unique Jewels", league.getName(), dataAccessService.findUniqueJewelsByLeague(league).size());
+        });
+  }
+}
